ui and build improvements
added some 3rd party libs today! man, cmake is pretty complicated,
so it took a while to figure out, but hey, it seems to be working,
so that's good!

I added two libraries:
- `tinyfiledialogs` to handle opening a file-select window (that
  varies per platform) in the case that no rom is specified on the
  commandline.
- `args` handles parsing and validation of commandline arguments.
  As I start to add more and more flags, it's probably a good idea
  to set up a more extensible framework to configure ANESE

Oh, and FYI, this is all me procrastinating working on the PPU.

weeeeeeee!! isn't emulation fun?!

+++ src/common/debug.h
@@ -6,6 +6,7 @@
 
 #include <cassert>
 #include <cstdio>
+#include <cstdlib>
 
 #include "interfaces/memory.h"
 #include "util.h"
@@ -51,7 +52,7 @@ public:
 
 // Some concrete singletons
 namespace Func_Memory_Functions {
-inline u8 give_me_rand() { return rand(); }
+inline u8 give_me_rand() { return u8(rand()); }
 inline u8 give_me_zero() { return 0x00; }
 inline u8 give_me_full() { return 0xFF; }
 }
